fix(web): revert move metadata to root layout
